# openapi/server.yaml
openapi: 3.0.3
info:
  title: "GISOPVK-Bot"
  description: "A chat bot to make registration and information access easier."
  version: 0.1.0
servers:
- url: http://{hostname}:{port}/api/
  variables:
    hostname:
      default: localhost
    port:
      default: "8080"
security:
  - bearerAuth: []

paths:
  /reply:
    post:
      operationId: getReply
      description: "Get a bot reply to a message."
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Query"
      responses:
        "200":
          description: "The request is valid, a bot reply is provided"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Reply"
        "400":
          description: "Some data in the request is invalid"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: "JWT token is not specified"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: "JWT token is invalid or expired"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    
  schemas:
    Query:
      type: object
      description: "Body of the GET request to the endpoint"
      properties:
        uuid:
          type: string
          format: uuid
          description: "User's unique identifier"
          example: "44d5ad73-576e-4c31-bf09-b53f5c0326ba"
        body:
          type: string
          description: "User message"
          example: "Как поступить в МГТУ им. Н. Э. Баумана?"
      required:
        - uuid
        - body

    Reply:
      type: object
      description: "Body of the GET response from the endpoint"
      properties:
        uuid:
          type: string
          format: uuid
          description: "User's unique identifier"
          example: "44d5ad73-576e-4c31-bf09-b53f5c0326ba"
        body:
          type: string
          description: "Bot reply"
          example: "Мужество, воля, труд и упорство!"
      required:
        - uuid
        - body

    Error:
      type: object
      description: "Error description"
      properties:
        message:
          type: string
          example: "Field \"uuid\" is missing"
      required:
        - message